Add support for for falling back through locale tags
Adds a static MessageFormat.Utils.getFallbackLocale function and an
instance property, fallbackLocale, which contains the locale of the
plural function being used.

When no plural function is found for the given locale
getFallbackLocale is used to try and find one by stepping through the
locale tags. This is done by removing the last tag from the locale and
checking for a plural function until there are no tags left. The tags
can be separated by either "-" (dash) or "_" (underscore).

For example:
```
  en_US          -> no match
  en             -> match

  zh-yue-Hant-HK -> no match
  zh-yue-Hant    -> no match
  zh-yue         -> no match
  zh             -> match (if locale/zh.js has been loaded)
```

If no match can be found then an Error is thrown.

fallbackLocale is used by precompile (and hence compile) to set the
plural function in the compiled output. This means that if you compile
a message when MessageFormat has a locale of "en-GB" but only the "en"
plural function is available, then the compiled message will always
use the "en" plural function, even if an "en-GB" plural function
becomes available later.