Fail-safe for no translation
This change simply passes through the original key in case there is no translation provided.

Before this change, if some text was untranslated, `{"Text goes here":""}` it would simply show up blank when used. This change makes it return the text in the original language, so the resulting UI is at least usable, albeit has elements displayed in the 'wrong' language.