Enhance publishing with Makefile
Add new targets *make parser* and a *make publish* in the Makefile.

- __make parser__ verify the presence of pegjs then call it to compile
the rules specified in *./lib/message_parser.pegjs* to the
*./lib/message_parser.js* file

- __make publish__ depends on the target described above, and generate
the ./messageformat.js root file, required by browsers.

This chain is intended to save time and avoir any conflicted version
between the development stuff and the production build.