Use CommonJS
This commit removes the custom UMD wrapper and uses the CommonJS
module format. This gives us a few benefits:

- We can use a real require/exports (in the tests too!)
- Compatibility with module loaders like webpack, Browserify

Webpack is used to generate the browser test bundle and a UMD
build in the dist directory.

Note: This isn't working in browsers yet because the make-plural
dep also uses a weird module system.