Add EasyCLA instructions
This PR proposes some succinct "heads up / how-to" instructions for using the CLA. It also includes a gentle reminder about joining the Foundation itself, because the people who can sign a CLA and the people who can sign a membership agreement are likely the same people. :-)

Please feel free to tweak this language if necessary. This can land at the same time we enable the CLA bot.

Signed-off-by: Brian Warner <brian@bdwarner.com>